refactor(codegen-new): split adapter_symbols.rs (895) into a folder module (<500 each)
Addresses the last review item: adapter_symbols.rs was still >500 after the drain.
Converts it to a folder module under the 500-line rule:
- adapter_symbols/mod.rs   (105) — JitSymbol + sym() + jit_symbols() composition
  (list_a ++ list_b) + the drift-guard test.
- adapter_symbols/list_a.rs (322) — first half of the engine-direct symbol list.
- adapter_symbols/list_b.rs (~480) — second half + the internal/generator helper
  sub-lists and the Registry harvest.

Pure code move (split at a clean `sym(...)` entry boundary) + per-file import trim.
The drift-guard test (no null / no conflicting dupe) still passes over the
reassembled table.

Validation: full TS suite = 327 files / 856 tests passed, 18 illegal, 0
can't-resolve — IDENTICAL to baseline (the carve dropped/duplicated no symbol, a
risk the coverage assert alone can't catch, so verified end-to-end). Lib suite
793 passed / 0 failed. Gate: adapter_symbols off the >500 list.
