+++ crates/rts-codegen-new/src/adapter_symbols/list_b.rs
@@ -1,386 +1,19 @@
-//! Engine-direct JIT symbols + the Registry-harvest composition for the
-//! `JITBuilder`.
-//!
-//! [`jit_symbols`] assembles the full JIT symbol table from THREE sources:
-//!  1. **Registry harvest** (`front::run::registry::all_jit_symbols`) — every
-//!     genuine namespace/class MEMBER that carries a real `fn_ptr` (e.g. the
-//!     `string` ns via `fp_for`, io, math, Date/URL class methods). The bulk; it
-//!     stays in sync with the Registry automatically, no hand list.
-//!  2. **`__rtsadp_*` adapter trampolines** — the engine's OWN codegen-owned
-//!     symbols (`crate::value::*`): the generic `+`, shape/IC obj access, function
-//!     values, dynamic dispatch. Not runtime, not in any Registry — listed here.
-//!  3. **Engine-internal runtime primitives** — `__RTS_FN_NS_GC_*` env / generator
-//!     / GEN_SM / string-pool ops the engine emits DIRECTLY as codegen sentinels
-//!     (NOT exposed as `rts:gc` members; the `gc` ns deliberately surfaces only
-//!     `collect`/`live_count`). The harvest cannot supply them (no member to hold
-//!     the address), so they are listed here too.
-//!
-//! Sources 2+3 are the irreducible "engine-direct" set — symbols the engine NAMES
-//! itself rather than resolving via the Registry. Source 1 used to be hand-listed
-//! here too; those entries drain out as each namespace's `register` is converted
-//! to carry real `fn_ptr`s (the `dataview`/`string` `fp_for` pattern).
-//!
-//! This replaces the fake `crate::runtime::symbols()` table. It is the new
-//! engine's analogue of the old engine's `register_runtime_symbols` /
-//! `runtime_jit_symbols` bridge (`rts-codegen-old/src/abi/mod.rs` +
-//! `codegen/jit.rs`) — but built against the `rts-runtime` FACADE, NOT by
-//! depending on the frozen old crate.
-//!
-//! ## Why sources 2+3 are listed by hand (the harvest cannot supply them)
-//!
-//! The harvest (`registry().jit_symbols()`) only yields members with a real,
-//! non-null `fn_ptr` (the **null-skip** invariant: alias/external members carry a
-//! null fn_ptr). The engine-internal `__RTS_FN_NS_GC_*` primitives (string-pool /
-//! env / generator / GEN_SM / gcell / poly-bridge) and a few directly-emitted
-//! class methods (`__RTS_FN_GL_STRING_*` slice/substr/codePointAt/localeCompare,
-//! emitted by `try_string_special`) are NOT registered as harvestable members —
-//! the engine NAMES them itself in its lowering — so there is no member to attach
-//! an address to. We take each one's address directly through the facade re-export
-//! and list it here. A genuine namespace/class MEMBER, by contrast, carries its
-//! real address at registration (`fp_for`) and is installed by the harvest with no
-//! hand entry (string ns / Date / URL already converted).
+//! Second half of the engine-direct JIT symbol list + the internal / generator
+//! helper sub-lists and the Registry harvest (split for the <500-line rule).
 
 use rts_engine::heap::env as rt_env;
-use rts_runtime::namespaces::gc::collector as rt_gcoll;
-use rts_runtime::namespaces::gc::handles as rt_handles;
-use rts_runtime::namespaces::gc::string_pool as rt_str;
 use rts_runtime::namespaces::globals::string::rt as rt_gl_str;
 use rts_runtime::namespaces::globals::proxy::ops as rt_proxy;
 
 use crate::value::{
-    abi_adapter, arraycb, arrayops, ctorval, dyndispatch, errslot, funcops, genops, genops_arith,
-    globalops, globalthis, inspect, iterops, objops, regexops,
+    arraycb, ctorval, dyndispatch, errslot, funcops, globalops, globalthis, iterops, objops,
+    regexops,
 };
 
-/// One installable JIT symbol: an extern "C" name and its function pointer. The
-/// pointer is to a `#[no_mangle] extern "C"` function with static lifetime.
-#[derive(Clone, Copy)]
-pub struct JitSymbol {
-    pub name: &'static str,
-    pub ptr: *const u8,
-}
+use super::{JitSymbol, sym};
 
-// SAFETY: every `ptr` is to static `extern "C"` code (the runtime binary or this
-// crate), never dereferenced as data — sound to share across threads.
-unsafe impl Send for JitSymbol {}
-unsafe impl Sync for JitSymbol {}
-
-/// The full set of `(symbol, fn_ptr)` pairs the JIT must install so the lowered
-/// code can `call` them by name: the REAL runtime symbols the new lowering emits
-/// today (gc string-pool, io, collections vec) + the codegen-owned `__rtsadp_*`
-/// adapter trampolines.
-///
-/// Every entry has a non-null pointer (the null-skip invariant from the engine's
-/// registry is satisfied by construction — we list only real bodies).
-pub fn jit_symbols() -> Vec<JitSymbol> {
-    let mut syms = vec![
-        // ---- module-level mutable global cells (epic #195) ----
-        sym(
-            "__RTS_FN_NS_GC_GCELL_GET",
-            rt_gcoll::__RTS_FN_NS_GC_GCELL_GET as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_GCELL_SET",
-            rt_gcoll::__RTS_FN_NS_GC_GCELL_SET as *const u8,
-        ),
-        // ---- REAL string pool (rts-std collector::string_pool) ----
-        sym(
-            "__RTS_FN_NS_GC_STRING_NEW",
-            rt_str::__RTS_FN_NS_GC_STRING_NEW as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_STRING_FROM_STATIC",
-            rt_str::__RTS_FN_NS_GC_STRING_FROM_STATIC as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_STRING_PTR",
-            rt_str::__RTS_FN_NS_GC_STRING_PTR as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_STRING_LEN",
-            rt_str::__RTS_FN_NS_GC_STRING_LEN as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_STRING_FREE",
-            rt_str::__RTS_FN_NS_GC_STRING_FREE as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_STRING_CONCAT",
-            rt_str::__RTS_FN_NS_GC_STRING_CONCAT as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_STRING_EQ",
-            rt_str::__RTS_FN_NS_GC_STRING_EQ as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_STRING_CMP",
-            rt_str::__RTS_FN_NS_GC_STRING_CMP as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_STRING_FROM_I64",
-            rt_str::__RTS_FN_NS_GC_STRING_FROM_I64 as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_STRING_FROM_F64",
-            rt_str::__RTS_FN_NS_GC_STRING_FROM_F64 as *const u8,
-        ),
-        // Replaces the old `__rtsadp_store/_load` indirection table: the payload
-        // carries the bare 48-bit slot+shard and the generation is reconstructed
-        // on demand from the live slot, so there is no side table to GC-root.
-        sym(
-            "__RTS_FN_NS_GC_POLY_FROM_HANDLE",
-            rt_handles::__RTS_FN_NS_GC_POLY_FROM_HANDLE as *const u8,
-        ),
-        sym(
-            "__RTS_FN_NS_GC_POLY_TO_HANDLE",
-            rt_handles::__RTS_FN_NS_GC_POLY_TO_HANDLE as *const u8,
-        ),
-        // ---- codegen-owned adapter trampolines (__rtsadp_*) ----
-        sym("__rtsadp_add", genops::__rtsadp_add as *const u8),
-        sym(
-            "__rtsadp_strict_eq",
-            genops::__rtsadp_strict_eq as *const u8,
-        ),
-        sym(
-            "__rtsadp_same_value",
-            genops::__rtsadp_same_value as *const u8,
-        ),
-        sym(
-            "__rtsadp_strict_neq",
-            genops::__rtsadp_strict_neq as *const u8,
-        ),
-        sym("__rtsadp_loose_eq", genops::__rtsadp_loose_eq as *const u8),
-        sym(
-            "__rtsadp_loose_neq",
-            genops::__rtsadp_loose_neq as *const u8,
-        ),
-        sym("__rtsadp_typeof", genops::__rtsadp_typeof as *const u8),
-        sym(
-            "__rtsadp_to_string",
-            genops::__rtsadp_to_string as *const u8,
-        ),
-        sym(
-            "__rtsadp_to_boolean",
-            genops::__rtsadp_to_boolean as *const u8,
-        ),
-        sym(
-            "__rtsadp_print_line",
-            abi_adapter::__rtsadp_print_line as *const u8,
-        ),
-        sym("__rtsadp_inspect", inspect::__rtsadp_inspect as *const u8),
-        sym(
-            "__rtsadp_inspect_object",
-            inspect::__rtsadp_inspect_object as *const u8,
-        ),
-        // ---- generic arithmetic / comparison / unary / bitwise (P4.8) ----
-        sym("__rtsadp_sub", genops_arith::__rtsadp_sub as *const u8),
-        sym("__rtsadp_mul", genops_arith::__rtsadp_mul as *const u8),
-        sym("__rtsadp_div", genops_arith::__rtsadp_div as *const u8),
-        sym("__rtsadp_mod", genops_arith::__rtsadp_mod as *const u8),
-        sym("__rtsadp_pow", genops_arith::__rtsadp_pow as *const u8),
-        sym("__rtsadp_lt", genops_arith::__rtsadp_lt as *const u8),
-        sym("__rtsadp_le", genops_arith::__rtsadp_le as *const u8),
-        sym("__rtsadp_gt", genops_arith::__rtsadp_gt as *const u8),
-        sym("__rtsadp_ge", genops_arith::__rtsadp_ge as *const u8),
-        sym("__rtsadp_neg", genops_arith::__rtsadp_neg as *const u8),
-        sym("__rtsadp_pos", genops_arith::__rtsadp_pos as *const u8),
-        sym("__rtsadp_bnot", genops_arith::__rtsadp_bnot as *const u8),
-        sym("__rtsadp_not", genops_arith::__rtsadp_not as *const u8),
-        sym("__rtsadp_band", genops_arith::__rtsadp_band as *const u8),
-        sym("__rtsadp_bor", genops_arith::__rtsadp_bor as *const u8),
-        sym("__rtsadp_bxor", genops_arith::__rtsadp_bxor as *const u8),
-        sym("__rtsadp_shl", genops_arith::__rtsadp_shl as *const u8),
-        sym("__rtsadp_shr", genops_arith::__rtsadp_shr as *const u8),
-        sym("__rtsadp_ushr", genops_arith::__rtsadp_ushr as *const u8),
-        // ---- codegen-owned Array trampolines (__rtsadp_arr_*, P4.5) ----
-        sym(
-            "__rtsadp_arr_index_of",
-            arrayops::__rtsadp_arr_index_of as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_includes",
-            arrayops::__rtsadp_arr_includes as *const u8,
-        ),
-        sym("__rtsadp_arr_at", arrayops::__rtsadp_arr_at as *const u8),
-        sym(
-            "__rtsadp_arr_join",
-            arrayops::__rtsadp_arr_join as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_join0",
-            arrayops::__rtsadp_arr_join0 as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_slice0",
-            arrayops::__rtsadp_arr_slice0 as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_to_string",
-            arrayops::__rtsadp_arr_to_string as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_push",
-            arrayops::__rtsadp_arr_push as *const u8,
-        ),
-        sym("__rtsadp_arr_pop", arrayops::__rtsadp_arr_pop as *const u8),
-        sym(
-            "__rtsadp_arr_slice",
-            arrayops::__rtsadp_arr_slice as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_last_index_of",
-            arrayops::__rtsadp_arr_last_index_of as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_reverse",
-            arrayops::__rtsadp_arr_reverse as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_fill",
-            arrayops::__rtsadp_arr_fill as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_concat",
-            arrayops::__rtsadp_arr_concat as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_splice",
-            arrayops::__rtsadp_arr_splice as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_to_spliced_var",
-            arrayops::__rtsadp_arr_to_spliced_var as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_flat",
-            arrayops::__rtsadp_arr_flat as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_shift",
-            arrayops::__rtsadp_arr_shift as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_unshift",
-            arrayops::__rtsadp_arr_unshift as *const u8,
-        ),
-        // ES2023 / arity-variant Array trampolines
-        sym("__rtsadp_arr_slice1", arrayops::__rtsadp_arr_slice1 as *const u8),
-        sym(
-            "__rtsadp_arr_index_of_from",
-            arrayops::__rtsadp_arr_index_of_from as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_includes_from",
-            arrayops::__rtsadp_arr_includes_from as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_last_index_of_from",
-            arrayops::__rtsadp_arr_last_index_of_from as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_to_reversed",
-            arrayops::__rtsadp_arr_to_reversed as *const u8,
-        ),
-        sym("__rtsadp_arr_with", arrayops::__rtsadp_arr_with as *const u8),
-        sym(
-            "__rtsadp_arr_flat_depth",
-            arrayops::__rtsadp_arr_flat_depth as *const u8,
-        ),
-        sym("__rtsadp_arr_sort", arrayops::__rtsadp_arr_sort as *const u8),
-        sym(
-            "__rtsadp_arr_sort_cmp",
-            arrayops::__rtsadp_arr_sort_cmp as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_to_sorted",
-            arrayops::__rtsadp_arr_to_sorted as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_to_sorted_cmp",
-            arrayops::__rtsadp_arr_to_sorted_cmp as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_to_spliced",
-            arrayops::__rtsadp_arr_to_spliced as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_copy_within",
-            arrayops::__rtsadp_arr_copy_within as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_copy_within2",
-            arrayops::__rtsadp_arr_copy_within2 as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_copy_within1",
-            arrayops::__rtsadp_arr_copy_within1 as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_fill2",
-            arrayops::__rtsadp_arr_fill2 as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_fill3",
-            arrayops::__rtsadp_arr_fill3 as *const u8,
-        ),
-        // ---- codegen-owned GLOBAL constant/function + Array/String STATIC
-        //      trampolines (__rtsadp_g_* / __rtsadp_arr_* / __rtsadp_str_*, P5.2) ----
-        sym(
-            "__rtsadp_g_number",
-            globalops::__rtsadp_g_number as *const u8,
-        ),
-        sym(
-            "__rtsadp_g_string",
-            globalops::__rtsadp_g_string as *const u8,
-        ),
-        sym(
-            "__rtsadp_g_boolean",
-            globalops::__rtsadp_g_boolean as *const u8,
-        ),
-        sym(
-            "__rtsadp_g_parse_int",
-            globalops::__rtsadp_g_parse_int as *const u8,
-        ),
-        sym(
-            "__rtsadp_g_parse_float",
-            globalops::__rtsadp_g_parse_float as *const u8,
-        ),
-        sym(
-            "__rtsadp_g_is_nan",
-            globalops::__rtsadp_g_is_nan as *const u8,
-        ),
-        sym(
-            "__rtsadp_g_is_finite",
-            globalops::__rtsadp_g_is_finite as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_is_array",
-            globalops::__rtsadp_arr_is_array as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_new_sized",
-            globalops::__rtsadp_arr_new_sized as *const u8,
-        ),
-        sym(
-            "__rtsadp_arr_from",
-            globalops::__rtsadp_arr_from as *const u8,
-        ),
-        sym(
-            "__rtsadp_str_from_char_code",
-            globalops::__rtsadp_str_from_char_code as *const u8,
-        ),
-        sym(
-            "__rtsadp_str_from_char_code_arr",
-            globalops::__rtsadp_str_from_char_code_arr as *const u8,
-        ),
-        sym(
-            "__rtsadp_str_from_code_point",
-            globalops::__rtsadp_str_from_code_point as *const u8,
-        ),
-        sym(
-            "__rtsadp_str_split",
-            globalops::__rtsadp_str_split as *const u8,
-        ),
+pub(super) fn symbols() -> Vec<JitSymbol> {
+    let mut v: Vec<JitSymbol> = vec![
         sym(
             "__rtsadp_math_reduce",
             globalops::__rtsadp_math_reduce as *const u8,
@@ -685,24 +318,13 @@ pub fn jit_symbols() -> Vec<JitSymbol> {
             dyndispatch::__rtsadp_dyn_repeat as *const u8,
         ),
     ];
-    syms.extend(gl_method_symbols());
-    syms.extend(test_framework_symbols());
-    syms.extend(gc_internal_symbols());
-    syms.extend(generator_symbols());
-    // Pilar 6: the REAL `__RTS_FN_GL_DATE_*` / `__RTS_FN_NS_DATE_*` symbols the
-    // Registry-driven Date dispatch ([`crate::front::run::registry_call`]) emits
-    // directly — replacing the `__rtsadp_date_*` trampolines that used to forward
-    // to them.
-    syms
+    v.extend(gl_method_symbols());
+    v.extend(test_framework_symbols());
+    v.extend(gc_internal_symbols());
+    v.extend(generator_symbols());
+    v
 }
 
-
-/// The REAL global-class instance-method symbols the P4 data-driven dispatch
-/// ([`crate::dispatch`]) resolves and the method lowering emits — String + Number
-/// instance methods. Each is the ACTUAL `__RTS_FN_GL_*` extern (taken by address
-/// through the facade); the dispatch metadata references these exact names, so
-/// the symbol set is EXACTLY what the lowering can emit (a referenced-but-missing
-/// symbol would be the SIGILL-class bug we avoid).
 fn gl_method_symbols() -> Vec<JitSymbol> {
     vec![
         // String class methods the engine's OWN lowering emits DIRECTLY
@@ -854,43 +476,4 @@ fn generator_symbols() -> Vec<JitSymbol> {
         sym("__RTS_FN_NS_GC_ITER_VALUE", rt_gen::__RTS_FN_NS_GC_ITER_VALUE as *const u8),
         sym("__RTS_FN_NS_GC_ITER_DONE", rt_gen::__RTS_FN_NS_GC_ITER_DONE as *const u8),
     ]
-}
-
-#[inline]
-fn sym(name: &'static str, ptr: *const u8) -> JitSymbol {
-    JitSymbol { name, ptr }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::jit_symbols;
-    use std::collections::HashMap;
-
-    /// Drift/sanity guard for the JIT symbol table. NOTE: this checks the INSTALLED
-    /// set, not full emittable-symbol coverage — it cannot catch a symbol the
-    /// lowering can emit that is neither harvested nor hand-listed (that surfaces
-    /// only at runtime / in the TS suite as a "can't resolve symbol"). What it DOES
-    /// guarantee: every installed symbol has a REAL address (no null slipped through
-    /// from an `external` member whose `fp_for` forgot it — the link-OK/runtime-
-    /// SIGILL class), and no name is installed with TWO different addresses (a
-    /// harvest entry disagreeing with a hand entry; same-address dupes are harmless).
-    #[test]
-    fn jit_symbols_have_no_null_and_no_conflicting_dupes() {
-        let syms = jit_symbols();
-        assert!(!syms.is_empty(), "the JIT symbol table is empty");
-        let mut seen: HashMap<&str, *const u8> = HashMap::new();
-        for s in &syms {
-            assert!(!s.ptr.is_null(), "JIT symbol `{}` has a NULL fn_ptr", s.name);
-            match seen.get(s.name) {
-                Some(&prev) => assert_eq!(
-                    prev, s.ptr,
-                    "JIT symbol `{}` installed with two different addresses",
-                    s.name
-                ),
-                None => {
-                    seen.insert(s.name, s.ptr);
-                }
-            }
-        }
-    }
 }
